- updating to vs26 - fixing general problems with scheduler

+++ Engine/Gumball/Source/Gumball/Timer.cpp
@@ -9,8 +9,8 @@ void Timer::start() {
     time = high_resolution_clock::now();
 }
 float Timer::seconds() const {
-    return duration_cast<std::chrono::seconds>(high_resolution_clock::now() - time).count(); 
+    return (float)duration_cast<std::chrono::seconds>(high_resolution_clock::now() - time).count(); 
 }
 double Timer::milliseconds() const { 
-    return duration_cast<std::chrono::milliseconds>(high_resolution_clock::now() - time).count(); 
+    return (float)duration_cast<std::chrono::milliseconds>(high_resolution_clock::now() - time).count();
 }
